+++ sixtyfps_runtime/corelib/graphics.rs
@@ -96,19 +96,31 @@ pub type RenderingCache<T> = vec_arena::Arena<CachedGraphicsData<T>>;
 /// FontRequest collects all the developer-configurable properties for fonts, such as family, weight, etc.
 /// It is submitted as a request to the platform font system (i.e. CoreText on macOS) and in exchange the
 /// backend returns a Box<dyn Font>.
-#[derive(Debug, Clone, PartialEq)]
-#[repr(C)]
+#[derive(Debug, Clone, PartialEq, Default)]
 pub struct FontRequest {
     /// The name of the font family to be used, such as "Helvetica". An empty family name means the system
     /// default font family should be used.
-    pub family: SharedString,
+    pub family: Option<SharedString>,
     /// If the weight is None, the the system default font weight should be used.
     pub weight: Option<i32>,
     /// If the pixel size is None, the system default font size should be used.
     pub pixel_size: Option<f32>,
     /// The additional spacing (or shrinking if negative) between glyphs. This is usually not submitted to
     /// the font-subsystem but collected here for API convenience
-    pub letter_spacing: f32,
+    pub letter_spacing: Option<f32>,
+}
+
+impl FontRequest {
+    /// Consumes the FontRequest, replaces any missing fields from the specified other request and
+    /// returns the new request.
+    pub fn merge(self, other: &FontRequest) -> Self {
+        Self {
+            family: self.family.or_else(|| other.family.clone()),
+            weight: self.weight.or_else(|| other.weight.clone()),
+            pixel_size: self.pixel_size.or_else(|| other.pixel_size.clone()),
+            letter_spacing: self.letter_spacing.or_else(|| other.letter_spacing.clone()),
+        }
+    }
 }
 
 /// The FontMetrics trait is constructed from a FontRequest by the graphics backend and supplied to text related